Split in memory implementation to separate files.
